pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.1",
    "bs4>=0.0.2",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "transformers>=5.3.0",
]
